requirements.txt
numpy
opencv-python
easyocr
openpyxl
streamlit
pdf2image
Pillow
pypdfium2
pypdf2
pandas
